Update in propa toolbox : clean and remove deprecated
