+++ propa/kraken_toolbox/params.py
@@ -1,0 +1,2 @@
+# Path to the kraken exe file (kraken.exe) for Windows only. For Linux, the kraken executable does not need to be copied to a specific directory.
+KRAKEN_BIN_DIRECTORY = r"C:\Users\baptiste.menetrier\Desktop\ressource\AcousticToolbox\cygwinfield_x_20201102kraken"
