+++ propa/kraken_toolbox/post_process.py
@@ -12,7 +12,7 @@ from scipy.interpolate import interp1d
 
 
 def postprocess_ir(shd_fpath, source, rcv_range, rcv_depth):
-    """Post process Kraken run to derive ocean waveguide impulse response."""
+    """Read kraken results to get the transfert function."""
 
     # Dummy read to get frequencies used by kraken and field grid information
     _, _, freqVec, _, _, _, field_pos, _ = readshd(filename=shd_fpath, freq=0)
@@ -59,10 +59,9 @@ def postprocess_received_signal(
     propagating_spectrum = source.positive_spectrum[source.positive_freq > fc]
 
     # TODO : check which factor to apply
-    # norm_factor = (
-    #     RHO_W / np.pi * np.exp(1j * (2 * np.pi * positive_fft_freq / C0 - np.pi / 2))
-    # )
-    norm_factor = 1
+    norm_factor = (
+        RHO_W / np.pi * np.exp(1j * (2 * np.pi * propagating_freq / C0 - np.pi / 2))
+    )
 
     # Received signal spectrum resulting from the convolution of the source signal and the impulse response
     transmited_field_f = mult_along_axis(
@@ -139,7 +138,7 @@ def postprocess_ir_from_broadband_pressure_field(
     fc = waveguide_cutoff_freq(waveguide_depth=minimum_waveguide_depth)
     propagating_freq = source.positive_freq[source.positive_freq > fc]
 
-    # # TODO need to be fixed
+    # # TODO interpolation to match source frequencies not implemented yet
     # if (
     #     frequencies.size < propagating_freq.size
     # ):  # Sparse frequency vector used for kraken
@@ -258,116 +257,6 @@ def interp_frequency_domain_ir(pf_p, f_p, f):
             )
 
     return interpolated_pf
-
-
-def process_broadband(fname, source, max_depth):
-    """Write frequencies to env file and run KRAKEN"""
-
-    if source.kraken_freq is None:
-        kraken_freq = source.positive_freq
-    else:
-        kraken_freq = source.kraken_freq
-
-    # Get rid of frequencies below the cutoff frequency
-    fc = C0 / (4 * max_depth)
-    idx_freq = kraken_freq > fc
-    kraken_freq = kraken_freq[idx_freq]
-    # source_spectrum = source_spectrum[idx_freq]
-
-    # Write env file with frequencies of interest
-    with open(fname + ".env", "r") as f:
-        lines = f.readlines()
-        lines[-2] = f"{int(len(kraken_freq))}              ! Nfreq\n"
-        kraken_freqvec = " ".join([str(f) for f in kraken_freq])
-        lines[-1] = f" {kraken_freqvec} /      ! freqVec( 1 : Nfreq )"
-
-    with open(fname + ".env", "w") as f:
-        f.writelines(lines)
-
-    # Run kraken for the frequencies of interest
-    runkraken(fname)  # TODO: deprecated (needs env flp and freqs as args)
-
-
-def fourier_synthesis_kraken(fname, source, max_depth):
-    # OUTDATED functin to be removed
-    """Derive time serie of the received signal through Fourier synthesis.
-
-    Note that according to Jensen et al. (2000), the sampling frequency must be greater than 8 * fmax for visual inspection of the propagated pulse.
-    """
-
-    # TODO: add args to define the frequencies to use
-    # It can be very usefull to reduce the number of frequencies to compute the Fourier synthesis
-    # For instance using a non unigorm frequency vector can help catching the entire signal complexity while significantly reducing the
-    # computing effort --> example : signal with dirac like harmonics
-
-    # freq_ = source.energetic_freq[source.energetic_freq > 0]  # Option to reduce computing effort (need to update time vector accordingly)
-
-    positive_fft_freq = source.freq[source.freq > 0]
-    if source.kraken_freq is None:
-        kraken_freq = source.freq[source.freq > 0]
-    else:
-        kraken_freq = source.kraken_freq
-        # source_spectrum = np.interp(freq_, source.freq, source.spectrum)
-
-    positive_source_spectrum = source.spectrum[source.freq > 0]
-
-    # Get rid of frequencies below the cutoff frequency
-    fc = C0 / (4 * max_depth)
-    idx_freq = kraken_freq > fc
-    kraken_freq = kraken_freq[idx_freq]
-    # source_spectrum = source_spectrum[idx_freq]
-
-    # Write env file with frequencies of interest
-    # TODO: update section below with appropriate write_env function ?
-    with open(fname + ".env", "r") as f:
-        lines = f.readlines()
-        lines[-2] = f"{int(len(kraken_freq))}              ! Nfreq\n"
-        kraken_freqvec = " ".join([str(f) for f in kraken_freq])
-        lines[-1] = f" {kraken_freqvec} /      ! freqVec( 1 : Nfreq )"
-
-    with open(fname + ".env", "w") as f:
-        f.writelines(lines)
-
-    # Run kraken for the frequencies of interest
-    runkraken(fname)  # TODO: deprecated (needs env flp and freqs as args)
-
-    env_filename = fname + ".shd"
-
-    rcv_signal_f = []
-    for ifreq, freq in enumerate(positive_fft_freq):  # Iterate over all fft frequencies
-        # Load pressure field for closest freq
-        _, _, _, _, fread, _, field_pos, pressure = readshd(
-            filename=env_filename, freq=freq
-        )
-
-        # Normalisation factor to get potential velocity
-        norm_factor = 4 * RHO_W * np.exp(1j * (2 * np.pi * freq / C0 - np.pi / 2))
-
-        # Potential velocity field for freq f <-> spatial transfert function
-        p_f = np.squeeze(pressure, axis=(0, 1)) * norm_factor
-
-        # Received signal in the frequency domain
-        s_f = positive_source_spectrum[ifreq] * p_f
-        # s_f = source.spectrum[ifreq] * p_f
-
-        # Set approximated time of arrival
-        # c_group_max = 1500
-        # if ifreq == 0:
-        #     d_direct_path = np.sqrt(rcv_range**2 + (rcv_depth - Pos["s"]["z"]) ** 2)
-        #     tau = d_direct_path / c_group_max
-        # s_f *= np.exp(1j * 2 * np.pi * tau * freq)  # Apply delay to the signal
-
-        rcv_signal_f.append(s_f)
-
-    received_signal_f = np.array(rcv_signal_f)
-
-    # real inverse FFT to exploit conjugate symmetry of the transfert function (see Jensen et al. 2000 p.612-613)
-    received_signal_t = np.fft.irfft(received_signal_f, axis=0)
-
-    transmited_field = np.real(received_signal_t)
-    time_vector = np.arange(0, transmited_field.shape[0] / source.fs, 1 / source.fs)
-
-    return time_vector, transmited_field, field_pos
 
 
 if __name__ == "__main__":
