+++ propa/kraken_toolbox/tests/kraken_env/bathy_data.csv
@@ -1,0 +1,3 @@
+  0.0, 3000
+  4, 3000
+  10, 2500
