--- propa/kraken_toolbox/usefull_path.py
@@ -1,1 +1,0 @@
-KRAKEN_BIN_DIRECTORY = r"C:\Users\baptiste.menetrier\Desktop\ressource\AcousticToolbox\cygwinfield_x_20201102kraken"
